Change list-remove icon to a minus sign
It seems to be a convention across several icons themes to use a minus
sign for the list-remove icon. The icon themes that do this include, but
are not limited to: Oxygen, Gnome, Adwaita, FaenzaFlattr, HighContrast.
Especially since Oxygen used such a symbol for list-remove, it seems
appropriate for Breeze to do so as well.

Furthermore, a minus sign feels (at least to me) to be more appropriate
for list-remove, rather than a circle-backslash symbol/interdictory
circle, as I find it more clearly symbolises removal.

This icon was made from the list-add icon, by removing the vertical
rectangle, and colouring it appropriately according to the Breeze colour
guidelines for destructive actions.